### 2023-05-28 #5423 @devarops
> wave @rashatwi @devarops @ritika-giri would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sure, @osorensen, I can review it.
