### 2023-05-29 #5496 @editorialbot
[REVIEW]: iTensor: An R package for independent component analysis-based matrix/tensor decomposition
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/iTensor
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
