### 2023-06-12 rikenbit/iTensor#4 @devarops
[JOSS REVIEW] Test coverage is insufficient
@kokitsuyuzaki:

The current test coverage for the code is approximately 35.8%, which falls short of providing sufficient verification of the software's functionality and justifiably recommending its acceptance.

My suggestion is to strive for a minimum coverage of 50% in each file or, ideally, attain a target of 90% coverage throughout the entire repository. Through this approach, we can considerably enhance the reliability and robustness of the software.
