+++ tests/testthat.R
@@ -5,6 +5,8 @@ library(rTensor)
 options(testthat.use_colours = FALSE)
 
 # Basic usage
+test_file("testthat/test-toyModel.R")
+test_file("testthat/test-CorrIndex.R")
 test_file("testthat/test-ICA.R")
 test_file("testthat/test-ICA2.R")
 test_file("testthat/test-MICA.R")
