+++ tests/testthat/test-GroupICA.R
@@ -1,74 +1,146 @@
 # Simulation Datasets
 d <- 6
-X1 <- matrix(runif(200 * d), nrow = 200, ncol = d)
-X2 <- matrix(runif(150 * d), nrow = 150, ncol = d)
-X3 <- matrix(runif(250 * d), nrow = 250, ncol = d)
+X1 <- matrix(runif(20 * d), nrow = 20, ncol = d)
+X2 <- matrix(runif(15 * d), nrow = 15, ncol = d)
+X3 <- matrix(runif(25 * d), nrow = 25, ncol = d)
 
 Xs <- list(X1 = X1, X2 = X2, X3 = X3)
 
 ## Perform GroupICA against Simulation Datasets
 J1 <- d
-out.pooled <- GroupICA(Xs, J1=J1, algorithm="pooled", ica.algorithm = "FastICA")
-# out.Calhoun2009 <- GroupICA(Xs, J1=J1, algorithm="Calhoun2009", ica.algorithm = "FastICA")
-# out.Pfister2018 <- GroupICA(Xs, J1=J1, algorithm="Pfister2018", ica.algorithm = "FastICA")
+set.seed(12345)
+out.pooled <- GroupICA(Xs, J1=J1, algorithm="pooled", ica.algorithm = "FastICA", verbose=TRUE)
+set.seed(12345)
+out.Calhoun2009 <- GroupICA(Xs, J1=J1, algorithm="Calhoun2009", ica.algorithm = "FastICA", verbose=TRUE)
+set.seed(12345)
+out.Pfister2018 <- suppressWarnings(GroupICA(Xs, J1=J1, algorithm="Pfister2018", ica.algorithm = "FastICA", verbose=TRUE))
+
+set.seed(12345)
+out.pooled_2 <- GroupICA(Xs, J1=J1, algorithm="pooled", ica.algorithm = "JADE", verbose=TRUE)
+set.seed(12345)
+out.Calhoun2009_2 <- GroupICA(Xs, J1=J1, algorithm="Calhoun2009", ica.algorithm = "JADE", verbose=TRUE)
+set.seed(12345)
+out.Pfister2018_2 <- suppressWarnings(GroupICA(Xs, J1=J1, algorithm="Pfister2018", ica.algorithm = "JADE", verbose=TRUE))
+
+# Test Hidden Functions
+set.seed(12345)
+out.pooled_ <- iTensor:::.pooled(Xs, J1, J2=J1,
+    ica.algorithm="FastICA",
+    num.iter=30, thr=1E-10, verbose=FALSE)
+set.seed(12345)
+out.Calhoun2009_ <- iTensor:::.Calhoun2009(Xs, J1, J2=J1,
+    ica.algorithm="FastICA",
+    num.iter=30, thr=1E-10, verbose=FALSE)
+set.seed(12345)
+out.Pfister2018_ <- suppressWarnings(iTensor:::.Pfister2018(Xs, J1, J2=J1,
+    ica.algorithm="FastICA",
+    num.iter=30, thr=1E-10, verbose=FALSE))
+expect_identical(out.pooled, out.pooled_)
+expect_identical(out.Calhoun2009, out.Calhoun2009_)
+expect_identical(out.Pfister2018, out.Pfister2018_)
+expect_true(is.numeric(iTensor:::.eachidx(J1, 1, 1)))
+expect_true(is.numeric(iTensor:::.generate_subgroups(1, 1)))
 
 ## Test Output object / type
 ### Test O-1: Object
 expect_identical(is.list(out.pooled), TRUE)
-# expect_identical(is.list(out.Calhoun2009), TRUE)
[…]
