### 2023-06-13 rikenbit/iTensor#4 @devarops
Hello @kokitsuyuzaki,

I noticed that there is a failing test in line 41 of `test-MICA.R`. Please make the necessary modifications to ensure that all tests pass.

Once all tests pass, please let me know so I can continue with my review.

![image](https://github.com/rikenbit/iTensor/assets/3461018/afdb21ea-581d-4d78-8a2b-ab50a13eb645)
