+++ tests/testthat/test-MICA.R
@@ -38,7 +38,7 @@ expect_identical(out$J, J)
 expect_identical(out$eta, 1000*1e-4)
 
 ### Test O-9: verbose
-expect_identical(out$verbose, formals(MICA)$verbose)
+expect_identical(!out$verbose, formals(MICA)$verbose)
 
 ### Test O-10: RecError
 expect_identical(is.vector(out$ABChange), TRUE)
