### 2023-06-14 rikenbit/iTensor#4 @kokitsuyuzaki
Thanks, I modified the `test-MICA.R`.
https://github.com/rikenbit/iTensor/commit/72001ed7defe6341ce4411ec348f5a1432f2d791
