### 2023-06-22 #5496 @osorensen
@devarops, is [this issue](https://github.com/rikenbit/iTensor/issues/4) addressed in a satisfactory way? If so, could you please update your checklist? Otherwise, please elaborate what the authors need to do further. Thank you.
