### 2023-06-23 #5496 @devarops
@osorensen:
Yes, the authors have adequately resolved the issue. I recommend to **accept** this submission.
