### 2023-06-23 rikenbit/iTensor#6 @osorensen
Missing DOIs in JOSS submission
This is my first editorial comment on your JOSS submission.

All DOIs seem to be missing from the references, please add all of them to the `doi` key in your BibTeX source:

```

OK DOIs

- None

MISSING DOIs

- 10.7551/mitpress/7011.003.0009 may be a valid DOI for title: An information-maximization approach to blind separation and blind deconvolution
- 10.1162/089976699300016719 may be a valid DOI for title: Independent component analysis using an extended infomax algorithm for mixed subgaussian and supergaussian sources
- 10.1109/72.761722 may be a valid DOI for title: Fast and robust fixed-point algorithms for independent component analysis
- 10.1049/ip-f-2.1993.0054 may be a valid DOI for title: Blind beamforming for non-gaussian signals
- 10.1007/978-3-642-15995-4_21 may be a valid DOI for title: Auxiliary-Function-Based Independent Component Analysis for Super-Gaussian Sources
- 10.1186/1471-2105-13-24 may be a valid DOI for title: Independent Principal Component Analysis for biologically meaningful dimension reduction of large biological data sets
- 10.1109/31.76486 may be a valid DOI for title: Indeterminacy and identifiability of blind identification
- 10.1109/icassp.1989.266878 may be a valid DOI for title: Source separation using higher order moments
- 10.1109/ijcnn.1999.831077 may be a valid DOI for title: MICA: Multimodal independent component analysis
- 10.1016/j.neuroimage.2008.10.057 may be a valid DOI for title: A review of group ICA for fMRI data and ICA for joint inference of imaging, genetic, and ERP data
- 10.1016/j.sigpro.2022.108457 may be a valid DOI for title: CorrIndex: a permutation invariant performance index
- 10.1109/memb.2006.1607672 may be a valid DOI for title: Unmixing fMRI with Independent Component Analysis
- 10.1038/nbt.2859 may be a valid DOI for title: The dynamics and regulators of cell fate decisions are revealed by pseudotemporal ordering of single cells
- 10.3389/fnins.2013.00267 may be a valid DOI for title: MEG and EEG data analysis with MNE-Python
- 10.1371/journal.pone.0270556 may be a valid DOI for title: Comparing the reliability of different ICA algorithms for fMRI analysis
- 10.1109/spmb.2016.7846858 may be a valid DOI for title: Applying fully tensorial ICA to fMRI data

INVALID DOIs

- None
```

https://github.com/openjournals/joss-reviews/issues/5496
