### 2023-06-23 rikenbit/iTensor#7 @osorensen
Editorial comment
Please change from "Of" to "of" on line 19:

https://github.com/openjournals/joss-reviews/issues/5496
