### 2023-06-23 rikenbit/iTensor#8 @osorensen
Editorial comment
Please make sure the word "Figure" appears only once on line 35:

https://github.com/openjournals/joss-reviews/issues/5496
