### 2023-06-23 rikenbit/iTensor#9 @osorensen
Editorial comment
Please remove the "first author" footnote, as you are the only author:

https://github.com/rikenbit/iTensor/blob/72001ed7defe6341ce4411ec348f5a1432f2d791/paper/paper.md?plain=1#L11

https://github.com/openjournals/joss-reviews/issues/5496
