### 2023-06-23 rikenbit/iTensor#10 @osorensen
Editorial comment
Given that you load the package `iTensor`, the qualifier `iTensor::` in this line is redundant. I suggest changing it to `data1 <- toyModel("ICA_Type1")`.

https://github.com/rikenbit/iTensor/blob/72001ed7defe6341ce4411ec348f5a1432f2d791/paper/paper.md?plain=1#L54

https://github.com/openjournals/joss-reviews/issues/5496
