### 2023-06-23 3d7a8c9
Added period at the end of figure caption.
