### 2023-06-25 49c8762
paper and bib modified
