+++ paper/paper.bib
@@ -9,7 +9,8 @@
   title = {An information-maximization approach to blind separation and blind deconvolution},
   volume = {7(6)},
   pages = {1129-1159},
-  year = {1995}}
+  year = {1995},
+  doi = {10.7551/mitpress/7011.003.0009}}
 
 @article{infomax2,
   author = {Amari, S. et al.},
@@ -24,7 +25,8 @@
   title = {Independent component analysis using an extended infomax algorithm for mixed subgaussian and supergaussian sources},
   volume = {11(2)},
   pages = {417-441},
-  year = {1999}}
+  year = {1999},
+  doi = {10.1162/089976699300016719}}
 
 % FastICA
 @article{fastica,
@@ -33,7 +35,8 @@
   title = {Fast and robust fixed-point algorithms for independent component analysis},
   volume = {10(3)},
   pages = {626-634},
-  year = {1999}}
+  year = {1999},
+  doi = {10.1109/72.761722}}
 
 % JADE
 @article{jade,
@@ -42,7 +45,8 @@
   title = {Blind beamforming for non-gaussian signals},
   volume = {140(6)},
   pages = {362-370},
-  year = {1993}}
+  year = {1993},
+  doi = {10.1049/ip-f-2.1993.0054}}
 
 % AuxICA1/2
 @article{auxica,
@@ -51,7 +55,8 @@
   title = {Auxiliary-Function-Based Independent Component Analysis for Super-Gaussian Sources},
   volume = {6365},
   pages = {165-172},
-  year = {2010}}
+  year = {2010},
+  doi = {10.1007/978-3-642-15995-4_21}}
 
 % IPCA
 @article{ipca,
@@ -59,7 +64,8 @@
   journal = {BMC Bioinformatics},
   title = {Independent Principal Component Analysis for biologically meaningful dimension reduction of large biological data sets},
   volume = {13(24)},
-  year = {2012}}
+  year = {2012},
+  doi = {10.1186/1471-2105-13-24}}
 
 % SIMBEC
 @article{simbec,
@@ -76,7 +82,8 @@
   title = {Indeterminacy and identifiability of blind identification},
   volume = {38(5)},
   pages = {499-509},
-  year = {1991}}
+  year = {1991},
+  doi = {10.1109/31.76486}}
 
 % SOBI
 @article{sobi,
@@ -94,7 +101,8 @@
   title = {Source separation using higher order moments},
   volume = {4},
   pages = {2109-2112},
-  year = {1989}}
+  year = {1989},
+  doi = {10.1109/icassp.1989.266878}}
 
 % ProDenICA
 @article{prodenica,
@@ -117,7 +125,8 @@
   title = {MICA: Multimodal independent component analysis},
   volume = {2},
   pages = {927-932},
-  year = {1999}}
+  year = {1999},
+  doi = {10.1109/ijcnn.1999.831077}}
 
 % GroupICA
 @article{groupica1,
@@ -126,7 +135,8 @@
   title = {A review of group ICA for fMRI data and ICA for joint inference of imaging, genetic, and ERP data},
   volume = {45(1 Suppl)},
   pages = {S163-S172},
-  year = {2009}}
+  year = {2009},
+  doi = {10.1016/j.neuroimage.2008.10.057}}
 
 @article{groupica2,
   author = {Pfister, N. et al.},
@@ -148,7 +158,8 @@
   title = {CorrIndex: a permutation invariant performance index},
   volume = {195},
   pages = {108457},
-  year = {2022}}
+  year = {2022},
+  doi = {10.1016/j.sigpro.2022.108457}}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
[…]
