### 2023-06-25 rikenbit/iTensor#6 @kokitsuyuzaki
Thanks for spotting the typo.
I added all the DOIs  https://github.com/rikenbit/iTensor/commit/49c876261ef7a548977f7bd7a568475b6aa67d20
