### 2023-06-25 #5496 @kokitsuyuzaki
@osorensen
Thank you for the suggestions.
I modified the manuscript so please check it.
https://github.com/rikenbit/iTensor/commit/49c876261ef7a548977f7bd7a568475b6aa67d20
