### 2023-06-26 #5496 @kokitsuyuzaki
I tagged the latest version as v1.0.3 and archieved by Zenodo.

The DOI is 10.5281/zenodo.8080600

Please also check these pages.
https://github.com/rikenbit/iTensor/releases/tag/v1.0.3
https://doi.org/10.5281/zenodo.8080600
