### 2023-06-26 #5496 @osorensen
Thanks @kokitsuyuzaki.

- [ ] Could you please change the title of the Zenodo archive from "rikenbit/iTensor: iTensor: An R package for independent component analysis-based matrix/tensor decomposition" to "iTensor: An R package for independent component analysis-based matrix/tensor decomposition", so it exactly matches the paper?

- [ ] Also, it is not necessary to include the name of me or any reviewers in the author list in the Zenodo archive. Could you please remove them? If you want to thank a review, please add an acknowledgement at the end of the paper instead.

Once done, please report back to me in this thread.
