### 2023-06-26 da1afb8
Update paper.bib
Protect word "Gaussian" so it begins with an uppercase letter.
