+++ paper/paper.bib
@@ -39,7 +39,7 @@
 @article{jade,
   author = {Cardoso, J.-F},
   journal = {IEEE Proceedings F},
-  title = {Blind beamforming for non-gaussian signals},
+  title = {Blind beamforming for {non-Gaussian} signals},
   volume = {140(6)},
   pages = {362-370},
   year = {1993}}
