### 2023-06-26 #5496 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
