### 2025-08-02 #5496 @gkthiruvathukal
@openjournals/joss-eics  Please investigate this. I don't even see any command issued by @kumeS. We've seen folks posting or interacting with review issues after closure and need to find a solution as we all have enough workload as is.
